Refactor Incident Details mapping (#684)
* Refactor alert handler

* update tests

* fix TS errors

* remove extra line

* fix ts errors
