+++ frontend_vue/src/components/tokens/pwa/ActivatedToken.vue
@@ -23,7 +23,6 @@ const props = defineProps<{
 }>();
 
 defineEmits(['howToUse']);
-console.log(props.tokenData);
 
 const tokenSnippetData = ref({
   url: props.tokenData.token_url || '',
